+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Added validation and formatting for the number field.
+* Cancel the editing state of a fields when the escape key is pressed.
 * The next field is now selected when the tab character is pressed when a field is
   selected.
 * Changed the styling of the notification alerts.